Create English_PremierLeague_Position.py
Here I wrote a code to determine the league positions of clubs in the English Premier League.

How it works:

Step 1: A football fan will be asked to enter the name of a preferred club.

Step 2: The code will run to check the club's position whether a Top 4 or not in Top 4.

Step 3: The club's position on the premier league table will be displayed otherwise, the response will be that the club is not in top 4.

Step 4: The date in which it was last updated will be posted.

Note: Does anyone know how I can write a code that update the league standing automatically otherwise I may have to update this at the end of the 21/22 season.